fix ui model bug
