+++ cocos/3d/framework/model-component.ts
@@ -193,6 +193,7 @@ export class ModelComponent extends RenderableComponent {
     public recreateModel () {
         if (this._model) {
             this._model.destroy();
+            this._model.scene.destroyModel(this._model);
             this._model = null;
         }
         this._updateModels();
