+++ cocos/3d/ui/components/ui-render-component.ts
@@ -232,21 +232,7 @@ export class UIRenderComponent extends UIComponent {
     }
 
     public onEnable () {
-        let parent = this.node;
-        // 获取被渲染相机的 visibility
-        while (parent) {
-            if (parent) {
-                const canvasComp = parent.getComponent(CanvasComponent);
-                if (canvasComp) {
-                    this._visibility = canvasComp.visibility;
-                    break;
-                }
-            }
-
-            // @ts-ignore
-            parent = parent.parent;
-        }
-
+        super.onEnable();
         this.node.on(SystemEventType.ANCHOR_CHANGED, this._nodeStateChange, this);
         this.node.on(SystemEventType.TRANSFORM_CHANGED, this._nodeStateChange, this);
         this.node.on(SystemEventType.SIZE_CHANGED, this._nodeStateChange, this);
@@ -259,7 +245,7 @@ export class UIRenderComponent extends UIComponent {
     }
 
     public onDisable () {
-        this._visibility = -1;
+        super.onDisable();
         // this.node._renderComponent = null;
         // this.disableRender();
         this.node.off(SystemEventType.ANCHOR_CHANGED, this._nodeStateChange, this);
